tidy up test script
